Remove delay in dragging tabs
When one tries to drag a tab there is a lot of lag at the start of the
drag gesture. As a result the tab is likely to be trailing the mouse
pointer by inches so that figuring out where the tab will end up is
quite confusing.

BUG: 371939
Differential Revision: https://phabricator.kde.org/D7451

(cherry picked from commit 14bf3b418cc5f5f99f598b84a68f78f034b6676e)